# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRVM API Documentation",
        "version": "1.0.0",
        "description": "API documentation for MyReverseVendingMachine Server",
        "contact": {
            "email": "admin@myrvm.com"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server (Relative)"
        }
    ],
    "paths": {
...
        "/api/v1/tenant/vouchers": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "List all vouchers owned by authenticated tenant",
                "operationId": "getTenantVouchers",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vouchers"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Tenant Role Required"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}